requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
uagents>=0.17.0
